Rename files in connections folder
