+++ package.json
@@ -49,7 +49,7 @@
   "dependencies": {
     "@elastic/elasticsearch": "^7.4.0",
     "@polkadot/types": "1.28.1",
-    "@subsocial/api": "latest",
+    "@subsocial/api": "0.4.10",
     "@subsocial/types": "0.4.8-beta3",
     "@subsocial/utils": "^0.1.23",
     "@types/request-promise": "^4.1.46",
