+++ src/connections/elastic.ts
@@ -1,16 +1,18 @@
 import { Client } from '@elastic/elasticsearch'
 import { elasticLog as log } from './loggers'
 
-require('dotenv').config();
+require('dotenv').config()
+
+const esUrl = process.env.ES_NODE_URL
 
 let greeted = false
 if (!greeted) {
   greeted = true
-  log.info(`Connecting to Elasticsearch at ${process.env.ES_NODE_URL}`)
+  log.info(`Connecting to Elasticsearch at ${esUrl}`)
 }
 
 export const elasticIndexer = new Client({
-  node: process.env.ES_NODE_URL,
+  node: esUrl,
   auth: {
     username: process.env.ES_OFFCHAIN_USER,
     password: process.env.ES_OFFCHAIN_PASSWORD
@@ -18,7 +20,7 @@ export const elasticIndexer = new Client({
 })
 
 export const elasticReader = new Client({
-  node: process.env.ES_NODE_URL,
+  node: esUrl,
   auth: {
     username: process.env.ES_READONLY_USER,
     password: process.env.ES_READONLY_PASSWORD
