+++ src/postgres/delete-activity.ts
@@ -1,4 +1,4 @@
-import { pg } from '../connections/connect-postgres';
+import { pg } from '../connections/postgres';
 import { encodeStructId } from '../substrate/utils';
 import { PostId, SpaceId } from '@subsocial/types/substrate/interfaces/subsocial';
 import { deleteNotificationsLog, deleteNotificationsLogError } from './postges-logger';
