+++ src/postgres/event-handlers/SpaceFollowed.ts
@@ -1,5 +1,5 @@
 import { SpaceId } from '@subsocial/types/substrate/interfaces/subsocial';
-import { substrate } from '../../connections/connect-subsocial';
+import { substrate } from '../../connections/subsocial';
 import { insertSpaceFollower } from '../insert-follower';
 import { insertActivityForSpace } from '../insert-activity';
 import { insertNotificationForOwner } from '../notifications';
