+++ src/search/event-handlers/PostCreated.ts
@@ -1,4 +1,4 @@
-import { substrate } from '../../connections/connect-subsocial';
+import { substrate } from '../../connections/subsocial';
 import { EventHandlerFn } from '../../substrate/types';
 import { parsePostEvent } from '../../substrate/utils';
 import { indexPostContent } from './utils';
